requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
itsdangerous
email-validator 